[copilot_studio_skill] working version
