--- templates/azure-bot/src/api/teams_package/manifest.json
@@ -1,49 +1,0 @@
-{
-  "$schema": "https://developer.microsoft.com/en-us/json-schemas/teams/v1.19/MicrosoftTeams.schema.json",
-  "manifestVersion": "1.19",
-  "version": "1.0.0",
-  "id": "__teamsAppId",
-  "developer": {
-    "name": "Contoso",
-    "websiteUrl": "https://www.microsoft.com",
-    "privacyUrl": "https://www.microsoft.com",
-    "termsOfUseUrl": "https://www.microsoft.com"
-  },
-  "icons": {
-    "color": "color.png",
-    "outline": "outline.png"
-  },
-  "name": {
-    "short": "__teamsAppName",
-    "full": "full desc"
-  },
-  "description": {
-    "short": "short description",
-    "full": "full desc"
-  },
-  "bots": [
-    {
-      "botId": "__botAppId",
-      "scopes": ["personal", "team", "groupChat"],
-      "supportsFiles": false,
-      "isNotificationOnly": false,
-      "commandLists": [
-        {
-          "scopes": ["personal"],
-          "commands": [
-            {
-              "title": "How can you help me?",
-              "description": "How can you help me?"
-            }
-          ]
-        }
-      ]
-    }
-  ],
-  "accentColor": "#FFFFFF",
-  "composeExtensions": [],
-  "configurableTabs": [],
-  "staticTabs": [],
-  "permissions": ["identity", "messageTeamMembers"],
-  "validDomains": []
-}
