+++ templates/copilot_studio_skill/azure.yaml
