+++ templates/copilot_studio_skill/infra/aca.bicep
