+++ templates/copilot_studio_skill/infra/acr.bicep
