+++ templates/copilot_studio_skill/infra/appin.bicep
