+++ templates/copilot_studio_skill/infra/bot.bicep
