+++ templates/copilot_studio_skill/infra/cosmos.bicep
