+++ templates/copilot_studio_skill/infra/fetch-container-image.bicep
