+++ templates/copilot_studio_skill/infra/main.bicep
