+++ templates/copilot_studio_skill/infra/main.parameters.json
