+++ templates/copilot_studio_skill/infra/openAI.bicep
