+++ templates/copilot_studio_skill/infra/uami.bicep
