+++ templates/copilot_studio_skill/src/api/.dockerignore
