+++ templates/copilot_studio_skill/src/api/config.py
@@ -7,6 +7,9 @@ load_dotenv(override=True)
 class Config:
     """Bot Configuration"""
 
+    HOST = os.getenv("HOST", "localhost")
+    PORT = int(os.getenv("PORT", 8080))
+
     # DO NOT CHANGE THIS KEYS!!
     # These keys are used to validate the bot's identity
     # and must match these named as Bot configuration expects
@@ -30,6 +33,10 @@ class Config:
     AZURE_OPENAI_API_VERSION = os.getenv("AZURE_OPENAI_API_VERSION")
 
     def validate(self):
+        if not self.HOST or not self.PORT:
+            raise Exception(
+                "Missing required configuration. HOST and PORT must be set."
+            )
         if not self.APP_ID or not self.APP_PASSWORD or not self.APP_TENANTID:
             raise Exception(
                 "Missing required configuration. APP_ID, APP_PASSWORD, and APP_TENANT_ID must be set."
