+++ templates/copilot_studio_skill/src/api/sk_conversation_agent.py
